:recycle: encapsulated all scripts in a python packaged that can be found by setuptools.find_packages()
